business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetAntennaNr.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetAntennaNr implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+Y ");
        messageString.append(command.getParameter("Antenna"));
        messageString.append("\n");
        return messageString;
    }
}
